functions/requirements.txt
transformers==4.35.0
torch
google-cloud-storage
functions-framework
sentencepiece
